Nueva carpeta/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
pygame>=2.5.0
scipy>=1.10.0
pyttsx3>=2.90
gtts>=2.5.0

